+++ src/Extension/Twig/Node/ResolvePuliPathsNode.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * This file is part of the Twig Puli Extension.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Webmozart\Puli\Extension\Twig\Node;
+
+use Twig_Node;
+
+/**
+ * Marks templates loaded by {@link \Webmozart\Puli\Extension\Twig\Loader\PuliLoader}.
+ *
+ * @since  1.0
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+class ResolvePuliPathsNode extends Twig_Node
+{
+}