+++ src/Extension/Twig/PuliExtension.php
@@ -1,0 +1,69 @@
+<?php
+
+/*
+ * This file is part of the Twig Puli Extension.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Webmozart\Puli\Extension\Twig;
+
+use Twig_NodeVisitorInterface;
+use Webmozart\Puli\Locator\ResourceLocatorInterface;
+use Webmozart\Puli\Extension\Twig\NodeVisitor\RelativePathResolver;
+use Webmozart\Puli\Extension\Twig\TokenParser\ResolvePuliPathsTokenParser;
+
+/**
+ * @since  1.0
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+class PuliExtension extends \Twig_Extension
+{
+    const VERSION = '@package_version@';
+
+    const RELEASE_DATE = '@release_date@';
+
+    /**
+     * @var ResourceLocatorInterface
+     */
+    private $locator;
+
+    public function __construct(ResourceLocatorInterface $locator)
+    {
+        $this->locator = $locator;
+    }
+
+    /**
+     * Returns the name of the extension.
+     *
+     * @return string The extension name
+     */
+    public function getName()
+    {
+        return 'puli';
+    }
+
+    /**
+     * Returns the node visitor instances to add to the existing list.
+     *
+     * @return Twig_NodeVisitorInterface[] An array of Twig_NodeVisitorInterface instances
+     */
+    public function getNodeVisitors()
+    {
+        return array(new RelativePathResolver($this->locator));
+    }
+
+    /**
+     * Returns the token parser instances to add to the existing list.
+     *
+     * @return array An array of Twig_TokenParserInterface or Twig_TokenParserBrokerInterface instances
+     */
+    public function getTokenParsers()
+    {
+        return array(new ResolvePuliPathsTokenParser());
+    }
+
+}