+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Changelog
  * renamed `PhpResourceLocator` to `PhpCacheLocator`
  * renamed `PhpResourceLocatorDumper` to `PhpCacheDumper`
  * added `FilesystemLocator`
+ * removed `ResourceDiscoveringInterface`
+ * a base `ResourceLocatorInterface` can now be passed to `ResourceRepository`
+ * instead of arrays, `ResourceCollection` objects are now returned everywhere
+ * renamed `ResourceInterface::getPath()` to `getRealPath()`
+ * renamed `ResourceInterface::getRepositoryPath()` to `getPath()`
 
 * 1.0.0-alpha2 (2014-02-14)